Fixed bad relative paths for testing

+++ gocd/data_source_job_definition.go
@@ -148,11 +148,11 @@ func dataSourceGocdJobTemplateRead(d *schema.ResourceData, meta interface{}) err
 	}
 
 	if ric, ok := d.GetOk("run_instance_count"); ok {
-		j.RunInstanceCount = int64(ric.(int))
+		j.RunInstanceCount = ric.(int)
 	}
 
 	if to, ok := d.GetOk("timeout"); ok {
-		j.Timeout = int64(to.(int))
+		j.Timeout = to.(int)
 	}
 
 	if envVars, ok := d.Get("environment_variables").([]interface{}); ok && len(envVars) > 0 {
